+++ fltk-sys/src/group.rs
@@ -559,6 +559,18 @@ extern "C" {
     pub fn Fl_Group_set_current(grp: *mut Fl_Group);
 }
 extern "C" {
+    pub fn Fl_Group_draw_child(self_: *const Fl_Group, w: *mut Fl_Widget);
+}
+extern "C" {
+    pub fn Fl_Group_update_child(self_: *const Fl_Group, w: *mut Fl_Widget);
+}
+extern "C" {
+    pub fn Fl_Group_draw_outside_label(self_: *const Fl_Group, widget: *const Fl_Widget);
+}
+extern "C" {
+    pub fn Fl_Group_draw_children(self_: *mut Fl_Group);
+}
+extern "C" {
     pub fn Fl_Group_begin(self_: *mut Fl_Group);
 }
 extern "C" {
@@ -590,6 +602,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Group_resizable(self_: *mut Fl_Group, arg1: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Group_set_clip_children(self_: *mut Fl_Group, c: libc::c_int);
+}
+extern "C" {
+    pub fn Fl_Group_clip_children(self_: *mut Fl_Group) -> libc::c_int;
 }
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
@@ -895,6 +913,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Pack_resizable(self_: *mut Fl_Pack, arg1: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Pack_set_clip_children(self_: *mut Fl_Pack, c: libc::c_int);
+}
+extern "C" {
+    pub fn Fl_Pack_clip_children(self_: *mut Fl_Pack) -> libc::c_int;
 }
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
@@ -1219,6 +1243,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Scroll_resizable(self_: *mut Fl_Scroll, arg1: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Scroll_set_clip_children(self_: *mut Fl_Scroll, c: libc::c_int);
+}
+extern "C" {
+    pub fn Fl_Scroll_clip_children(self_: *mut Fl_Scroll) -> libc::c_int;
 }
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
@@ -1546,6 +1576,12 @@ extern "C" {
 extern "C" {
     pub fn Fl_Tabs_resizable(self_: *mut Fl_Tabs, arg1: *mut libc::c_void);
 }
+extern "C" {
+    pub fn Fl_Tabs_set_clip_children(self_: *mut Fl_Tabs, c: libc::c_int);
+}
+extern "C" {
+    pub fn Fl_Tabs_clip_children(self_: *mut Fl_Tabs) -> libc::c_int;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Tile {
@@ -1844,6 +1880,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Tile_resizable(self_: *mut Fl_Tile, arg1: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Tile_set_clip_children(self_: *mut Fl_Tile, c: libc::c_int);
+}
+extern "C" {
+    pub fn Fl_Tile_clip_children(self_: *mut Fl_Tile) -> libc::c_int;
 }
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
@@ -2159,6 +2201,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Wizard_resizable(self_: *mut Fl_Wizard, arg1: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Wizard_set_clip_children(self_: *mut Fl_Wizard, c: libc::c_int);
+}
+extern "C" {
+    pub fn Fl_Wizard_clip_children(self_: *mut Fl_Wizard) -> libc::c_int;
 }
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
@@ -2492,4 +2540,10 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Color_Chooser_resizable(self_: *mut Fl_Color_Chooser, arg1: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Color_Chooser_set_clip_children(self_: *mut Fl_Color_Chooser, c: libc::c_int);
+}
+extern "C" {
+    pub fn Fl_Color_Chooser_clip_children(self_: *mut Fl_Color_Chooser) -> libc::c_int;
 }